requirements.txt
amqp==5.1.1
anyjson==0.3.3
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2022.12.7
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dataclasses==0.8
Django==3.2.16
django-appconf==1.0.5
django-celery-beat==2.2.1
django-celery-email==3.0.0
django-timezone-field==4.2.3
djangorestframework==3.14.0
idna==3.4
importlib-metadata==4.8.3
kombu==5.1.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pymongo==4.1.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.7
requests==2.27.1
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
